Qrcode/requirements.txt
colorama==0.4.6
pillow==11.1.0
qrcode==8.0
